Remove the old way of using React's context
