+++ test/test-react.js
@@ -1,21 +1,40 @@
 const test = require('ava')
 const React = require('react')
 const ReactDOM = require('react-dom')
-const { ProvideAtom, ConnectAtom, connect } = require('../src/react')
+const createContext = require('../src/react/context')
 const testApp = require('./generic-app')
 
 const h = React.createElement
 
-test('usage with react', t => {
-  const app = testApp({ h, ProvideAtom, ConnectAtom, connect })
+test('usage with react\'s new context', t => {
+  const app = testApp({ h, createContext })
 
   app.render((App, atom, root) => {
     ReactDOM.render(
-      h(ProvideAtom, { atom },
+      h(app.Provider, {},
         h(App, {})
       )
       , root)
   })
 
   app.assert(t)
+
+  ReactDOM.render(null, app.root)
+})
+
+test('usage with react\'s new context with atom as prop', t => {
+  const { Provider, Consumer, connect } = createContext()
+  const app = testApp({ h, Consumer, connect })
+
+  app.render((App, atom, root) => {
+    ReactDOM.render(
+      h(Provider, { atom },
+        h(App, {})
+      )
+      , root)
+  })
+
+  app.assert(t)
+
+  ReactDOM.render(null, app.root)
 })
